// front/src/app/services/Api/model/matchStat.ts
/**
 * KickStream API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MatchStat { 
    totalVotes: number;
    homePercentage: number;
    drawPercentage: number;
    awayPercentage: number;
    voteEnabled: boolean;
}